requirements.txt
streamlit
psycopg2-binary
cryptography
python-dotenv
fastapi
uvicorn
pyjwt
requests 